Rename + merge resouce cluster routes
