+++ mantis-control-plane/mantis-control-plane-server/src/main/java/io/mantisrx/master/resourcecluster/proto/ResourceClusterAPIProto.java
@@ -17,6 +17,8 @@
 package io.mantisrx.master.resourcecluster.proto;
 
 import io.mantisrx.master.jobcluster.proto.BaseResponse;
+import io.mantisrx.shaded.com.fasterxml.jackson.annotation.JsonCreator;
+import io.mantisrx.shaded.com.fasterxml.jackson.annotation.JsonProperty;
 import java.util.List;
 import lombok.Builder;
 import lombok.Singular;
@@ -24,18 +26,23 @@ import lombok.Value;
 
 public class ResourceClusterAPIProto {
 
+    // TODO: @Value generated equals doesn't include base class fields.
     @Value
     public static class ListResourceClustersResponse extends BaseResponse {
 
         @Singular
         List<RegisteredResourceCluster> registeredResourceClusters;
 
+        /** [Note] The @JsonCreator + @JasonProperty is needed when using this class with mixed shaded/non-shaded Jackson.
+         * The new @Jacksonized annotation is currently not usable with shaded Jackson here.
+         */
         @Builder
+        @JsonCreator
         public ListResourceClustersResponse(
-                final long requestId,
-                final ResponseCode responseCode,
-                final String message,
-                final List<RegisteredResourceCluster> registeredResourceClusters) {
+                @JsonProperty("requestId") final long requestId,
+                @JsonProperty("responseCode") final ResponseCode responseCode,
+                @JsonProperty("message") final String message,
+                @JsonProperty("registeredResourceClusters") final List<RegisteredResourceCluster> registeredResourceClusters) {
             super(requestId, responseCode, message);
             this.registeredResourceClusters = registeredResourceClusters;
         }
@@ -45,6 +52,14 @@ public class ResourceClusterAPIProto {
         public static class RegisteredResourceCluster {
             String id;
             String version;
+
+            @JsonCreator
+            public RegisteredResourceCluster(
+                @JsonProperty("id") final String id,
+                @JsonProperty("version") final String version) {
+                this.id = id;
+                this.version = version;
+            }
         }
     }
 
@@ -54,11 +69,12 @@ public class ResourceClusterAPIProto {
         MantisResourceClusterSpec clusterSpec;
 
         @Builder
+        @JsonCreator
         public GetResourceClusterResponse(
-                final long requestId,
-                final ResponseCode responseCode,
-                final String message,
-                final MantisResourceClusterSpec clusterSpec) {
+            @JsonProperty("requestId") final long requestId,
+            @JsonProperty("responseCode") final ResponseCode responseCode,
+            @JsonProperty("message") final String message,
+            @JsonProperty("clusterSpec") final MantisResourceClusterSpec clusterSpec) {
             super(requestId, responseCode, message);
             this.clusterSpec = clusterSpec;
         }
@@ -67,10 +83,11 @@ public class ResourceClusterAPIProto {
     @Value
     public static class DeleteResourceClusterResponse extends BaseResponse {
         @Builder
+        @JsonCreator
         public DeleteResourceClusterResponse(
-            final long requestId,
-            final ResponseCode responseCode,
-            final String message) {
+            @JsonProperty("requestId") final long requestId,
+            @JsonProperty("responseCode") final ResponseCode responseCode,
+            @JsonProperty("message") final String message) {
             super(requestId, responseCode, message);
         }
     }
