+++ mantis-control-plane/mantis-control-plane-server/src/main/java/io/mantisrx/master/resourcecluster/proto/ScaleResourceResponse.java
@@ -17,6 +17,8 @@
 package io.mantisrx.master.resourcecluster.proto;
 
 import io.mantisrx.master.jobcluster.proto.BaseResponse;
+import io.mantisrx.shaded.com.fasterxml.jackson.annotation.JsonCreator;
+import io.mantisrx.shaded.com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Builder;
 import lombok.Value;
 
@@ -33,15 +35,16 @@ public class ScaleResourceResponse extends BaseResponse {
     int desireSize;
 
     @Builder
+    @JsonCreator
     public ScaleResourceResponse(
-            final long requestId,
-            final ResponseCode responseCode,
-            final String message,
-            final String clusterId,
-            String skuId,
-            String region,
-            MantisResourceClusterEnvType envType,
-            int desireSize) {
+            @JsonProperty("requestId") final long requestId,
+            @JsonProperty("responseCode") final ResponseCode responseCode,
+            @JsonProperty("message") final String message,
+            @JsonProperty("clusterId") final String clusterId,
+            @JsonProperty("skuId") String skuId,
+            @JsonProperty("region") String region,
+            @JsonProperty("envType") MantisResourceClusterEnvType envType,
+            @JsonProperty("desireSize") int desireSize) {
         super(requestId, responseCode, message);
         this.clusterId = clusterId;
         this.skuId = skuId;
